Weapon Projectiles, Effects & Modifiers implemented
Implemented the weapon mods, various effects life dots and added them to the spell bottle prefab, enemies also use the same projectile mods so we can very easily give enemies dot dmg on hit, etc.

+++ Assets/Scripts/checkpoint.cs
@@ -22,7 +22,7 @@ public class checkpoint : MonoBehaviour
             GameData data = new GameData
             {
                 souls = SoulManagement.souls,
-                playerHP = gameManager.instance.playerScript.HP,
+                playerHP = (int)gameManager.instance.playerScript.HP,
                 checkpointPosition = transform.position
             };
             SaveManager.SaveGame( data);
